+++ src/configuration/plannerConfigurations.ts
@@ -306,6 +306,143 @@ export class Popf implements planner.PlannerProvider {
     }
 }
 
+export class Lpg implements planner.PlannerProvider {
+
+    get kind(): planner.PlannerKind {
+        return { kind: 'lpg-td' };
+    }
+
+    getNewPlannerLabel(): string {
+        return '$(mortar-board) LPG-td';
+    }
+
+    async configurePlanner(previousConfiguration?: planner.PlannerConfiguration | undefined): Promise<planner.PlannerConfiguration | undefined> {
+
+        const filters = os.platform() === 'win32' ?
+            {
+                'LPG-td Executable': ['exe']
+            }
+            : undefined;
+
+        const defaultUri: Uri | undefined = !!(previousConfiguration?.path) ?
+            Uri.file(previousConfiguration.path) :
+            undefined;
+
+        const popfUri = await selectedFile(`Select LPG-td`, defaultUri, filters);
+        if (!popfUri) { return undefined; }
+
+        const newPlannerConfiguration: planner.PlannerConfiguration = {
+            kind: this.kind.kind,
+            canConfigure: true,
+            path: popfUri.fsPath,
+            syntax: '$(planner) -o $(domain) -f $(problem) $(options)',
+            title: 'LPG-td'
+        };
+
+        return newPlannerConfiguration;
+    }
+
+    getPlannerOptions(): planner.PlannerOption[] {
+        // see https://lpg.unibs.it/lpg/README-LPGTD
+        return [
+            {
+                option: "-speed", description: "finds a solution (of any quality) as quickly as possible"
+            }, {
+                option: "-quality", description: "slower than in speed mode, but the planner finds a solution with better quality"
+            }, {
+                label: "-n <max number of desired solutions>", option: "-n 3"
+            }, {
+                option: "-noout", description: "no output file is produced"
+            }, {
+                option: "-v off", description: "switches off the verbose output of LPG (the planner provides only essential information)"
+            }, {
+                label: "-search_steps <integer>",
+                option: "-search_steps 500", description: "Specifies the initial number of search steps after which the search is "
+                    + "restarted. After each search restart, this number is automatically "
+                    + "incremented by a factor of 1.1. The default initial value for -search_steps "
+                    + "is 500.  Note that for simple problems this value could be significantly "
+                    + "reduced, obtaining better performance."
+
+            }, {
+                label: "-restarts <integer>",
+                option: "-restarts 9", description: `maximum number of search restarts after which the search `
+                    + `is repeated for a certain number of times(see the "-repeat" parameter). `
+                    + `After each restart, the value of some dynamic parameters (e.g., number `
+                    + `of search step) is automatically changed. The default value of -restarts `
+                    + `is 9.`
+
+            }, {
+                option: "-repeats 5",
+                label: "-repeats <integer>", description: `maximum number of times (repeats) the local search is `
+                    + `repeated to find the first solution. If no solution has been found within `
+                    + `the specified number of repeats (and the CPU-time limit for the local `
+                    + `search has not been exceeded), the best-first search is activated. `
+                    + `Each time the local search is repeated, the dynamic settings of the `
+                    + `planner (e.g., the number of search steps) are set to their initial value. `
+                    + `When LPG-td is run in quality mode or incremental mode ("-n x", x > 1), `
+                    + `if a solution is found by the local search (under the specified CPU-time `
+                    + `limit and number of repeats), the local search is repeated until the `
+                    + `CPU-time limit of the local search is reached, or the number of solutions `
+                    + `specified by the "-n" parameter have been found. `
+                    + `The default value of -repeats is 5. `
+
+            }, {
+                option: "-noise 0.1",
+                label: "-noise <number between 0 and 1>", description: `Specifies the initial noise value for Walkplan. Such value is dynamically `
+                    + `modified during each restart using a technique described in in Gerevini, `
+                    + `Saetti, Serina "An Empirical Analysis of Some Heuristic Features for Local `
+                    + `Search in LPG", ICAPS'04. The default value of -noise is 0.1. `
+
+            }, {
+                option: "-maxnoise <number>", description: "Specifies the maximum noise value that can (automatically) be reached by "
+                    + "the dynamic noise.  "
+
+            }, {
+                option: "-static_noise 0.1",
+                label: "-static_noise", description: `Switches off the dynamic noise during each restart. The value of the  `
+                    + `noise is fixed. The default value is 0.1, and it can be changed by using `
+                    + `the "-noise" parameter. `
+
+            }, {
+                label: "-seed <integer>", option: "-seed 2004",
+                description: `Specifies the seed for the random number generator used by Walkplan  `
+                    + `(a stochastic local search procedure). By using the same seed number,  `
+                    + `it is possible to repeat identical runs. The output files containing the `
+                    + `solutions produced by LPG include the seed number used by the planner to `
+                    + `generate them. In the 4th IPC, we used "-seed 2004". `
+
+            }, {
+                option: "-lowmemory", description: `With this option, the mutex relations between actions are computed at `
+                    + `runtime (instead of being computed before searching). We recommend the `
+                    + `use of this option only for very large problems. `
+
+            }, {
+                label: "-cputime <sec>", option: "-cputime 1800",
+                description: "Specifies the maximum CPU-time (in seconds) after which termination of  "
+                    + "the planning process is forced. The default value is 1800 (30 minutes). "
+
+            }, {
+                label: "-cputime_localsearch <sec>", option: "-cputime_localsearch 1200",
+                description: `When all restarts of the local search have been performed without finding `
+                    + `a solution, LPG runs in a best-first search based on Joerg Hoffman's `
+                    + `implementation (FF package v2.3). This option specifies the maximum `
+                    + `CPU-time (in seconds) after which the best-first search starts.  `
+                    + `The default value is 1200 (20 minutes). `
+
+            }, {
+                option: "-nobestfirst", description: "With this option, LPG-td does not run best-first search. "
+
+            }, {
+                option: "-onlybestfirst", description: "Forces the immediate run of the best-first search (no local search is performed).  "
+
+            }, {
+                option: "-timesteps", description: "This option can be used in STRIPS domains to define the plan quality  "
+                    + "metric as number of (Graphplan) time steps. "
+            },
+        ];
+    }
+}
+
 
 // const node: QuickPickItem = {
 //     label: "$(file-code) Select a Node.js file..."